+++ tcapy_examples/gen/metric_calculation_examples.py
@@ -28,12 +28,12 @@ ask_mid_bp = 0.1
 
 
 def get_sample_data():
-    from tcapy.analysis.algos.benchmark import BenchmarkSpreadToMid
+    from tcapy.analysis.algos.benchmark import BenchmarkMarketSpreadToMid
     logger.info("About to load data for " + ticker)
 
     tca_request = TCARequest(start_date='01 May 2017', finish_date='15 May 2017', ticker=ticker, trade_data_store='ms_sql_server',
                              market_data_store=market_data_store,
-                             benchmark_calcs=[BenchmarkSpreadToMid(bid_mid_bp=bid_mid_bp, ask_mid_bp=ask_mid_bp)],
+                             benchmark_calcs=[BenchmarkMarketSpreadToMid(bid_mid_bp=bid_mid_bp, ask_mid_bp=ask_mid_bp)],
                              trade_order_mapping=['trade_df'], tca_type=tca_type)
 
     tca_engine = TCAEngineImpl()