+++ tcapy_examples/gen/time_series_ops_examples.py
@@ -14,7 +14,7 @@ import os
 
 from tcapy.analysis.tcaengine import TCAEngineImpl
 from tcapy.analysis.tcarequest import TCARequest
-from tcapy.analysis.algos.benchmark import BenchmarkMid
+from tcapy.analysis.algos.benchmark import BenchmarkMarketMid
 from tcapy.util.timeseries import TimeSeriesOps
 from tcapy.util.loggermanager import LoggerManager
 
@@ -52,13 +52,13 @@ def get_sample_data():
                                  market_data_store=os.path.join(folder, 'small_test_market_df.csv.gz'),
                                  trade_order_mapping={'trade_df': os.path.join(folder, 'small_test_trade_df.csv'),
                                                       'order_df': os.path.join(folder, 'small_test_order_df.csv')},
-                                 tca_type=tca_type, benchmark_calcs=BenchmarkMid())
+                                 tca_type=tca_type, benchmark_calcs=BenchmarkMarketMid())
     else:
         tca_request = TCARequest(start_date=start_date, finish_date=finish_date, ticker=ticker,
                                  trade_data_store=trade_data_source,
                                  reporting_currency='EUR',
                                  market_data_store=data_source,
-                                 trade_order_mapping=['trade_df'], tca_type=tca_type, benchmark_calcs=BenchmarkMid())
+                                 trade_order_mapping=['trade_df'], tca_type=tca_type, benchmark_calcs=BenchmarkMarketMid())
 
     tca_engine = TCAEngineImpl()